backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras requirements.in
#
asgiref==3.8.1
    # via django
blessed==1.20.0
    # via -r requirements.in
boto3==1.35.69
    # via
    #   -r requirements.in
    #   django-ses
    #   pycognito
botocore==1.35.69
    # via
    #   boto3
    #   s3transfer
brotlipy==0.7.0
    # via -r requirements.in
certifi==2024.8.30
    # via
    #   requests
    #   sentry-sdk
cffi==1.17.1
    # via
    #   brotlipy
    #   cryptography
charset-normalizer==3.4.0
    # via requests
croniter==3.0.4
    # via -r requirements.in
cryptography==43.0.3
    # via pyjwt
diff-match-patch==20241021
    # via django-import-export
django==4.2.16
    # via
    #   -r requirements.in
    #   django-admin-autocomplete-filter
    #   django-allauth
    #   django-appconf
    #   django-auditlog
    #   django-crispy-forms
    #   django-guardian
    #   django-import-export
    #   django-model-utils
    #   django-picklefield
    #   django-q2
    #   django-recaptcha
    #   django-ses
    #   django-storages
    #   django-tinymce
    #   sentry-sdk
django-admin-autocomplete-filter==0.7.1
    # via -r requirements.in
django-admin-rangefilter==0.13.2
    # via -r requirements.in
django-allauth==65.0.2
    # via -r requirements.in
django-appconf==1.0.6
    # via django-avatar
django-auditlog==3.0.0
    # via -r requirements.in
django-avatar==8.0.1
    # via -r requirements.in
django-crispy-forms==2.3
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-file-resubmit==0.5.2
    # via -r requirements.in
django-guardian==2.4.0
    # via -r requirements.in
django-impersonate==1.9.4
    # via -r requirements.in
django-import-export==4.1.1
    # via -r requirements.in
django-model-utils==5.0.0
    # via -r requirements.in
django-picklefield==3.2
    # via django-q2
django-q2==1.7.4
    # via -r requirements.in
django-recaptcha==4.0.0
    # via -r requirements.in
django-ses==4.2.0
    # via -r requirements.in
django-spurl==0.6.8
    # via -r requirements.in
django-storages==1.14.4
    # via -r requirements.in
django-tinymce==4.1.0
    # via -r requirements.in
dnspython==2.7.0
    # via django-avatar
envs==1.4
    # via pycognito
gevent==24.10.3
    # via -r requirements.in
greenlet==3.1.1
    # via gevent
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.2
    # via gunicorn
pillow==10.4.0
    # via
    #   -r requirements.in
    #   django-avatar
psutil==6.1.0
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pycognito==2024.5.1
    # via -r requirements.in
pycparser==2.22
    # via cffi
pyjwt==2.10.0
    # via
    #   django-allauth
    #   pycognito
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   croniter
    #   django-auditlog
pytz==2024.2
    # via croniter
requests==2.32.3
    # via
    #   -r requirements.in
    #   django-allauth
    #   pycognito
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via django-allauth
s3transfer==0.10.4
    # via boto3
sentry-sdk==2.17.0
    # via -r requirements.in
six==1.16.0
    # via
    #   blessed
    #   django-spurl
    #   python-dateutil
sqlparse==0.5.2
    # via django
tablib==3.5.0
    # via django-import-export
urllib3==2.2.3
    # via
    #   botocore
    #   requests
    #   sentry-sdk
urlobject==2.4.3
    # via django-spurl
wcwidth==0.2.13
    # via blessed
whitenoise==6.7.0
    # via -r requirements.in
zope-event==5.0
    # via gevent
zope-interface==7.1.1
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
